Implemented ghost and rucman collision
If the ghosts are invulnerable, both the ghosts and rucman are reset. If the ghosts are vulnerable, only the ghost is reset. The reason why it's checked twice is that after rucman moves, there's a chance that he collides with something. Then, after the ghosts moves, there's a chance that it collides with something. If collision is checked once, if rucman and a ghost are moving in opposite directions and are adjacent squares, there's a chance that they'll pass through each other